e2e_tts/src/api/requirements.txt
einops==0.5.0
fastapi==0.88.0
ffmpy==0.3.0
numba==0.48
pyyaml
pydub==0.25.1
requests==2.28.1
uvicorn==0.20.0
Unidecode==1.3.4